% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummax.R
\name{argmax_g}
\alias{argmax_g}
\title{Compute per-group arg-max}
\usage{
argmax_g(col, first_indices)
}
\arguments{
\item{col}{vector to  max}

\item{first_indices}{numeric index of first index in each group}
}
\value{
per-group arg-max
}
\description{
Compute per-group arg-max
}
\examples{

v <- c(2, 3, 1, 7, 5)
fi <- c(1, 1, 3, 3, 3)
argmax_g(v, fi)

}
