% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum.R
\name{cumsum_g}
\alias{cumsum_g}
\title{Compute per-group cumulative sums}
\usage{
cumsum_g(col, first_indices, na.rm = FALSE)
}
\arguments{
\item{col}{vector to cumulative sum}

\item{first_indices}{numeric index of first index in each group}

\item{na.rm}{locial if TRUE remove NAs (empty sum is zero).}
}
\value{
per-group cumulative sum
}
\description{
Compute per-group cumulative sums
}
\examples{

v <- c(2, 3, 1, NA, 5)
fi <- c(1, 1, 3, 3, 3)
cumsum_g(v, fi)

}
