% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum.R
\name{cumsum_g_no_NA}
\alias{cumsum_g_no_NA}
\title{Compute per-group cumulative sums}
\usage{
cumsum_g_no_NA(col, first_indices)
}
\arguments{
\item{col}{vector to cumulative sum}

\item{first_indices}{numeric index of first index in each group}
}
\value{
per-group cumulative sum
}
\description{
Compute per-group cumulative sums
}
\examples{

v <- c(2, 3, 1, 4, 5)
fi <- c(1, 1, 3, 3, 3)
cumsum_g_no_NA(v, fi)

}
