% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{FastBaseR}
\alias{FastBaseR}
\alias{FastBaseR-package}
\title{\code{FastBaseR}: Fast cumulative grouped operations in R}
\description{
Supplies fast implementations of running (or per-row, or cumulative) per-group operations in R,
without the use of C, C++, data.table, or dplyr.
}
