% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstInGroup.R
\name{last_in_each_group}
\alias{last_in_each_group}
\title{Build the indexes of last indexes for each group}
\usage{
last_in_each_group(first_indices)
}
\arguments{
\item{n}{first_indices first index of each item (in order).}

\item{vector}{of last indices}
}
\description{
Build the indexes of last indexes for each group
}
\examples{

fi <- c(1, 1, 3, 3, 3)
last_in_each_group(fi)

}
