% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConwayLife.R
\name{life_step}
\alias{life_step}
\title{Forward sep a square matrix one iteration of Conway's come of life.}
\usage{
life_step(d)
}
\arguments{
\item{d}{a square matrix logical values}
}
\value{
next step matrix
}
\description{
Assumes zero-padded on all sides.
}
\examples{

d <- matrix(data = FALSE, nrow = 10, ncol = 10)
glider_txt <- "
.O
..O
OOO
"
glider <- read_cells(glider_txt)
d <- write_mat_region(d, 5, 5, glider)
life_step(d)


}
\seealso{
\code{\link{read_cells}}, \code{\link{write_mat_region}}
}
