% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstInGroup.R
\name{mark_first_in_each_group}
\alias{mark_first_in_each_group}
\title{Mark the first row in each group of rows.}
\usage{
mark_first_in_each_group(data, group_cols)
}
\arguments{
\item{data}{a data.frame ordered with respect to groups}

\item{group_cols}{character names of the grouping columns}
}
\value{
numeric vector with index of first row in each group
}
\description{
Mark the first row in each group of rows.
}
\examples{

d <- wrapr::build_frame(
    "g", "v" |
    1  , 1L  |
    0  , 2L  |
    0  , 3L  |
    1  , 4L  |
    1  , 5L  )

d <- d[wrapr::orderv(d), , drop = FALSE]
d$fi <- mark_first_in_each_group(d, "g")

print(d)

}
