% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummax.R
\name{max_g}
\alias{max_g}
\title{Compute per-group maxs}
\usage{
max_g(col, first_indices, na.rm = FALSE)
}
\arguments{
\item{col}{vector to  max}

\item{first_indices}{numeric index of first index in each group}

\item{na.rm}{logical if TRUE remove NAs}
}
\value{
per-group max
}
\description{
Compute per-group maxs
}
\examples{

v <- c(2, 3, 1, 7, 5)
fi <- c(1, 1, 3, 3, 3)
max_g(v, fi)

}
