% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConwayLife.R
\name{read_cells}
\alias{read_cells}
\title{Read life text file format.}
\usage{
read_cells(txt)
}
\arguments{
\item{txt}{incoming text as a single string}
}
\value{
matrix
}
\description{
Format is: rows of non-empty lines of . and O with ! headers
}
\examples{

glider <- "
.O
..O
OOO
"
read_cells(glider)

}
\seealso{
\code{\link{life_step}}, \code{\link{write_mat_region}}
}
