% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum.R
\name{sum_g}
\alias{sum_g}
\title{Compute per-group  sums}
\usage{
sum_g(col, first_indices, na.rm = FALSE)
}
\arguments{
\item{col}{vector to  sum}

\item{first_indices}{numeric index of first index in each group}

\item{na.rm}{locial if TRUE remove NAs}
}
\value{
per-group sum
}
\description{
Compute per-group  sums
}
\examples{

v <- c(2, 3, 1, NA, 5)
fi <- c(1, 1, 3, 3, 3)
sum_g(v, fi)

}
