% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConwayLife.R
\name{write_mat_region}
\alias{write_mat_region}
\title{Write a region of a life board.}
\usage{
write_mat_region(d, i, j, p)
}
\arguments{
\item{d}{life board matrix}

\item{i}{row to start writing}

\item{j}{column to start writing}

\item{p}{pattern matrix}
}
\description{
Write a region of a life board.
}
\seealso{
\code{\link{life_step}}, \code{\link{read_cells}}
}
